requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
python-docx==1.1.2
pdfplumber==0.11.5
pandas==2.2.3
pillow==11.1.0
scikit-learn==1.6.1
numpy==2.2.1
gunicorn==21.2.0
